requirements.txt
pyrogram
tgcrypto
python-dotenv
pymongo
